src/requirements.txt
Django==2.0.1
psycopg2==2.7.3.2
graphene==2.0.1
graphene_django==2.0.0
django-environ==0.4.4
html2text==2018.1.9
